Fixing squid: S2184 Math operands should be cast before assignment

+++ core/math/src/main/java/org/arakhne/afc/math/geometry/d2/ai/PathShadow2ai.java
@@ -306,7 +306,7 @@ public class PathShadow2ai<B extends Rectangle2ai<?, ?, ?, ?, ?, B>> {
 		if (sx0 > shadowXmax && sx1 > shadowXmax) {
 			// The line is entirely at the right of the shadow
 			if (sy1 != sy0) {
-				final double alpha = (sx1 - sx0) / (sy1 - sy0);
+				final double alpha = (double) (sx1 - sx0) / (sy1 - sy0);
 				if (sy0 < sy1) {
 					if (sy0 <= shadowYmin) {
 						final int xintercept = (int) Math.round(sx0 + (shadowYmin - sy0) * alpha);
